refactor: global 패키지로 패키지 구조 개선
