+++ src/main/java/roomescape/auth/AuthenticationPrincipal.java
@@ -1,4 +1,4 @@
-package roomescape.global;
+package roomescape.auth;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
