+++ src/main/java/roomescape/global/exception/DuplicateSaveException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.global.exception;
 
 public class DuplicateSaveException extends IllegalRequestException {
