+++ src/main/java/roomescape/global/exception/IllegalRequestException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.global.exception;
 
 public class IllegalRequestException extends RuntimeException {
