+++ src/main/java/roomescape/global/exception/InvalidTokenException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.global.exception;
 
 public class InvalidTokenException extends IllegalRequestException {
