+++ src/main/java/roomescape/global/exception/NoSuchRecordException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.global.exception;
 
 public class NoSuchRecordException extends IllegalRequestException {
