+++ src/main/java/roomescape/global/exception/ReservationTimeNotHourlyUnitException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.global.exception;
 
 public class ReservationTimeNotHourlyUnitException extends IllegalRequestException {
