+++ src/main/java/roomescape/global/exception/WrongPasswordException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.global.exception;
 
 public class WrongPasswordException extends IllegalRequestException {
