+++ src/main/java/roomescape/reservation/presentation/ReservationController.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.global.AuthenticationPrincipal;
+import roomescape.auth.AuthenticationPrincipal;
 import roomescape.member.domain.Member;
 import roomescape.reservation.dto.ReservationAddRequest;
 import roomescape.reservation.dto.ReservationResponse;
