+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -3,9 +3,9 @@ package roomescape.reservation.service;
 import java.time.LocalDate;
 import java.util.List;
 import org.springframework.stereotype.Service;
-import roomescape.exception.DuplicateSaveException;
-import roomescape.exception.IllegalReservationDateException;
-import roomescape.exception.NoSuchRecordException;
+import roomescape.global.exception.DuplicateSaveException;
+import roomescape.global.exception.IllegalReservationDateException;
+import roomescape.global.exception.NoSuchRecordException;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationRepository;
 import roomescape.reservation.dto.ReservationAddRequest;
