+++ src/test/java/roomescape/auth/service/AuthServiceTest.java
@@ -15,9 +15,9 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import roomescape.auth.dto.LoginRequest;
 import roomescape.auth.infrastructure.JwtTokenProvider;
-import roomescape.exception.InvalidTokenException;
-import roomescape.exception.NoSuchRecordException;
-import roomescape.exception.WrongPasswordException;
+import roomescape.global.exception.InvalidTokenException;
+import roomescape.global.exception.NoSuchRecordException;
+import roomescape.global.exception.WrongPasswordException;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.MemberRepository;
