+++ src/test/java/roomescape/reservation/service/ReservationServiceTest.java
@@ -20,9 +20,9 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import roomescape.exception.DuplicateSaveException;
-import roomescape.exception.IllegalReservationDateException;
-import roomescape.exception.NoSuchRecordException;
+import roomescape.global.exception.DuplicateSaveException;
+import roomescape.global.exception.IllegalReservationDateException;
+import roomescape.global.exception.NoSuchRecordException;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationRepository;
 import roomescape.reservation.dto.ReservationResponse;
