+++ src/test/java/roomescape/time/domain/ReservationTimeTest.java
@@ -6,7 +6,7 @@ import java.time.LocalTime;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import roomescape.exception.ReservationTimeNotHourlyUnitException;
+import roomescape.global.exception.ReservationTimeNotHourlyUnitException;
 
 class ReservationTimeTest {
