+++ src/test/java/roomescape/time/service/ReservationTimeServiceTest.java
@@ -15,7 +15,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import roomescape.exception.DuplicateSaveException;
+import roomescape.global.exception.DuplicateSaveException;
 import roomescape.time.domain.ReservationTime;
 import roomescape.time.domain.ReservationTimeRepository;
 import roomescape.time.dto.ReservationTimeResponse;
